+++ R/subset-extract.R
@@ -92,8 +92,12 @@ NULL
   # for ht[1,] <- x, nargs() is still 4
 
   n_idx <- nargs() - 2L
+  i_was_missing <- FALSE
   if (n_idx == 1) {
-    if (missing(i)) i <- seq_len(ncol(x))
+    if (missing(i)) {
+      i <- seq_len(ncol(x))
+      i_was_missing <- TRUE
+    }
     j <- i
     i <- seq_len(nrow(x))
   }
@@ -148,7 +152,14 @@ NULL
   # the below changes the data appropriately and retains all attributes, including
   # class. As we have already dealt with new rows/cols, we still have a valid
   # huxtable - all attributes have the correct dimensions:
-  res <- NextMethod()
+  res <- if (i_was_missing) {
+    # [<-.data.frame is weird, I know no way to get it to
+    # accept our new i and j. This is a workaround.
+    `[<-.data.frame`(x, , , value)
+  } else {
+    NextMethod()
+  }
+
   if (is_hux(value)) {
     res <- replace_properties(res, ix, jx, value)
   }