+++ NEWS.md
@@ -9,6 +9,7 @@
 * Bugfix: failure in `to_md()` with recent versions of `stringi` package.
 * Bugfix: repeating a single row in a subset, like `ht[c(1, 1, 2, 3), ]`,
   was setting `colspan = 2` on the repeated row.
+* Bugfix: zero-argument subset replacement like `ht[] <- ...` wasn't working.
 
 
 # huxtable 5.3.0